extract usePrefillPrescriptions hook
