+++ packages/components/src/systems/DraftPrescriptions/DraftPrescriptionsProvider.tsx
@@ -1,36 +1,32 @@
 import {
   Accessor,
   createContext,
-  createEffect,
   createMemo,
   createSignal,
   JSXElement,
   Setter,
   useContext
 } from 'solid-js';
-import {
-  MutationCreatePrescriptionsArgs,
-  Prescription,
-  PrescriptionState
-} from '@photonhealth/sdk/dist/types';
-import { DraftPrescriptionSource, PhotonClient } from '@photonhealth/sdk';
+import { Prescription, PrescriptionState } from '@photonhealth/sdk/dist/types';
+import { DraftPrescriptionSource } from '@photonhealth/sdk';
 import { buildPrescriptionSnapshot } from '../../analytics/buildFieldSnapshot';
 import { usePrescribeEventDispatch } from '../PrescribeEventDispatchProvider';
 import { useRecentOrders } from '../RecentOrders';
 import { usePhotonClient } from '../SDKProvider';
 import {
   CreatePrescription,
-  CreatePrescriptions,
   CreatePrescriptionTemplate,
-  GetPrescription,
   UpdatePrescriptionStates
 } from '../../fetch';
 import triggerToast from '../../utils/toastTriggers';
-import { constructRxNotes, formatPatientWeight } from './utils/formatters';
+import { formatPatientWeight } from './utils/formatters';
+import { mapFormDataToPrescriptionInput } from './utils/mappers';
 import {
-  InitialPrescriptionInput,
-  initialPrescriptionInputSchema
-} from './utils/initialPrescriptionInputSchema';
+  type InitialPrescriptionsPrefill,
+  type PrescriptionOverrides,
+  type TemplateOverrides,
+  usePrefillPrescriptions
+} from './utils/usePrefillPrescriptions';
 
 export type DraftPrescriptionsContextType = {
   // values
@@ -66,39 +62,6 @@ interface DraftPrescriptionProviderProps {
   weightUnit?: string;
 }
 
-export type TemplateOverrides = {
-  [templateId: string]: {
-    externalId?: string;
-    dispenseQuantity?: number;
-    dispenseUnit?: string;
-    dispenseAsWritten?: boolean;
-    fillsAllowed?: number;
-    daysSupply?: number;
-    instructions?: string;
-    notes?: string;
-  };
-};
-
-export type PrescriptionOverrides = {
-  [prescriptionId: string]: {
-    externalId?: string;
-    treatmentId?: string;
-    dispenseQuantity?: number;
-    dispenseUnit?: string;
-    dispenseAsWritten?: boolean;
-    fillsAllowed?: number;
-    daysSupply?: number;
-    instructions?: string;
-    notes?: string;
-    // ISO string or date string in format ex. 2026-10-02
-    doNotFillBeforeDate?: string;
-  };
-};
-
-// At runtime `solid-element`'s customElement parses the `supervisor` HTML
-// attribute as JSON. Valid JSON → object; invalid JSON → the raw string.
-export type InitialPrescriptionsPrefill = Partial<InitialPrescriptionInput>[] | string;
-
 export type PrescriptionFormData = {
   id?: string;
   doNotFillBeforeDate?: string;
@@ -118,25 +81,6 @@ export type PrescriptionFormData = {
   diagnoseCodes: string[];
 };
 
-const mapFormDataToPrescriptionInput = (
-  prescription: PrescriptionFormData,
-  treatmentId: string,
-  patientId: string
-) => ({
-  externalId: prescription.externalId,
-  patientId: patientId,
-  treatmentId: treatmentId,
-  dispenseAsWritten: prescription.dispenseAsWritten,
-  dispenseQuantity: prescription.dispenseQuantity,
-  dispenseUnit: prescription.dispenseUnit,
-  fillsAllowed: prescription.fillsAllowed,
-  daysSupply: prescription.daysSupply,
-  instructions: prescription.instructions,
-  notes: prescription.notes,
-  doNotFillBeforeDate: prescription.doNotFillBeforeDate,
-  diagnoses: prescription.diagnoseCodes
-});
-
 export type TryCreatePrescriptionTemplateOptions = {
   addToTemplates?: boolean;
   templateName?: string;
@@ -151,109 +95,11 @@ function isTreatmentInDraftPrescriptions(
   return draftedPrescriptions.some((draft) => draft.treatment.id === treatmentId);
 }
 
-const transformPrefillsToPrescriptionInputs = async ({
-  client,
-  patientId,
-  templateIdsPrefill,
-  templateOverrides,
-  prescriptionIdsPrefill,
-  prescriptionOverrides,
-  initialPrescriptions,
-  rxNotesPrefill
-}: {
-  client: PhotonClient;
-  patientId: string;
-  templateIdsPrefill: string[];
-  templateOverrides: TemplateOverrides;
-  prescriptionIdsPrefill: string[];
-  prescriptionOverrides: PrescriptionOverrides;
-  initialPrescriptions?: InitialPrescriptionsPrefill;
-  rxNotesPrefill: string;
-}) => {
-  const rxToCreate: MutationCreatePrescriptionsArgs['prescriptions'] = [];
-
-  // Create prescriptions from template IDs with possible overrides
-  if (templateIdsPrefill.length > 0) {
-    const dedupedTemplateIds = Array.from(new Set(templateIdsPrefill));
-    const templatedCreateRxList = dedupedTemplateIds.map((templateId) => {
-      return {
-        ...templateOverrides?.[templateId],
-        patientId: patientId,
-        templateId,
-        notes: constructRxNotes(
-          null,
-          templateOverrides?.[templateId]?.notes || null,
-          rxNotesPrefill
-        )
-      };
-    });
-    rxToCreate.push(...templatedCreateRxList);
-  }
-
-  // Create prescriptions from prescription IDs with possible overrides
-  if (prescriptionIdsPrefill.length > 0) {
-    const fetchedPrescriptions = await Promise.all(
-      prescriptionIdsPrefill.map(async (prescriptionId: string) => {
-        const overrides = prescriptionOverrides?.[prescriptionId] || {};
-        const { data } = await client.apollo.query({
-          query: GetPrescription,
-          variables: { id: prescriptionId },
-          // request should throw if there are any errors
-          // retrieving the prescription
-          errorPolicy: 'none'
-        });
-        if (data?.prescription) {
-          const input = mapFormDataToPrescriptionInput(
-            {
-              ...data.prescription,
-              ...overrides,
-              notes: constructRxNotes(
-                data.prescription.notes || null,
-                overrides.notes || null,
-                rxNotesPrefill
-              ),
-              doNotFillBeforeDate: overrides.doNotFillBeforeDate
-                ? overrides.doNotFillBeforeDate.slice(0, 10)
-                : undefined
-            },
-            overrides.treatmentId || data.prescription.treatment.id,
-            patientId
-          );
-          return input;
-        }
-        return null;
-      })
-    );
-    rxToCreate.push(...fetchedPrescriptions.filter((rx) => !!rx));
-  }
-
-  // Create prescriptions from initialPrescriptions JSON array
-  if (initialPrescriptions) {
-    if (typeof initialPrescriptions === 'string') {
-      throw new Error('Invalid JSON passed to initialPrescriptions');
-    }
-    for (const draft of initialPrescriptions) {
-      const result = initialPrescriptionInputSchema.safeParse(draft);
-
-      if (result.error) {
-        throw new Error(result.error.issues.map((i) => i.message).join(', '));
-      }
-
-      rxToCreate.push({ ...result.data, patientId });
-    }
-  }
-
-  return rxToCreate;
-};
-
 export const DraftPrescriptionsProvider = (props: DraftPrescriptionProviderProps) => {
   const { dispatchDraftPrescriptionCreated, dispatchAnalyticsTrackEvent } =
     usePrescribeEventDispatch();
   const [, recentOrdersActions] = useRecentOrders();
   const client = usePhotonClient();
-  const [hasCreatedPrefillPrescriptions, setHasCreatedPrefillPrescriptions] =
-    createSignal<boolean>(false);
-  const [isLoadingPrefills, setIsLoadingPrefills] = createSignal<boolean>(false);
   const [draftPrescriptions, setDraftPrescriptions] = createSignal<Prescription[]>([]);
 
   const prescriptionIds = createMemo(() =>
@@ -271,63 +117,9 @@ export const DraftPrescriptionsProvider = (props: DraftPrescriptionProviderProps
     return notesPrefill || '';
   });
 
-  // Prefill new prescriptions based on templateIds or prescriptionIds when we get a patientId
-  createEffect(async () => {
-    if (
-      (props.templateIdsPrefill.length > 0 ||
-        props.prescriptionIdsPrefill.length > 0 ||
-        props.initialPrescriptions) &&
-      // must have a patientId
-      !!props.patientId &&
-      // must not have created prescriptions yet
-      !hasCreatedPrefillPrescriptions()
-    ) {
-      setHasCreatedPrefillPrescriptions(true);
-      setIsLoadingPrefills(true);
-
-      try {
-        const prefillPrescriptionInputs = await transformPrefillsToPrescriptionInputs({
-          client,
-          patientId: props.patientId,
-          templateIdsPrefill: props.templateIdsPrefill,
-          templateOverrides: props.templateOverrides,
-          prescriptionIdsPrefill: props.prescriptionIdsPrefill,
-          prescriptionOverrides: props.prescriptionOverrides,
-          initialPrescriptions: props.initialPrescriptions,
-          rxNotesPrefill: rxNotesPrefill()
-        });
-
-        if (!prefillPrescriptionInputs.length) {
-          return;
-        }
-
-        const res = await client.apollo.mutate({
-          mutation: CreatePrescriptions,
-          variables: { prescriptions: prefillPrescriptionInputs }
-        });
-        const newRxs = res.data.createPrescriptions as Prescription[];
-
-        if (newRxs) {
-          setDraftPrescriptions((prev) => [...prev, ...newRxs]);
-          newRxs.forEach((rx) => {
-            dispatchDraftPrescriptionCreated(rx);
-            dispatchAnalyticsTrackEvent('ctaClicked', {
-              name: 'Draft Prescription Added',
-              draftPrescriptionSource: 'prefill',
-              fields: buildPrescriptionSnapshot(rx)
-            });
-          });
-        }
-      } catch (error) {
-        console.error('Error while trying to create prescriptions from prefills', { error });
-        triggerToast({
-          status: 'error',
-          body: 'There was an issue creating prescriptions from prefills. Please check your configuration.'
-        });
-      } finally {
-        setIsLoadingPrefills(false);
-      }
-    }
+  const { isLoadingPrefills } = usePrefillPrescriptions(props, {
+    rxNotesPrefill,
+    setDraftPrescriptions
   });
 
   const tryCreatePrescription = async (
