+++ packages/components/src/systems/DraftPrescriptions/index.tsx
@@ -1,3 +1,4 @@
 export * from './DraftPrescriptionItem';
 export * from './DraftPrescriptionsProvider';
+export * from './utils/usePrefillPrescriptions';
 export * from './utils/mappers';
