+++ packages/components/src/systems/DraftPrescriptions/utils/mappers.ts
@@ -1,5 +1,24 @@
 import { Prescription } from '@photonhealth/sdk/dist/types';
-import { PrescriptionFormData } from '../DraftPrescriptionsProvider';
+import type { PrescriptionFormData } from '../DraftPrescriptionsProvider';
+
+export const mapFormDataToPrescriptionInput = (
+  prescription: PrescriptionFormData,
+  treatmentId: string,
+  patientId: string
+) => ({
+  externalId: prescription.externalId,
+  patientId: patientId,
+  treatmentId: treatmentId,
+  dispenseAsWritten: prescription.dispenseAsWritten,
+  dispenseQuantity: prescription.dispenseQuantity,
+  dispenseUnit: prescription.dispenseUnit,
+  fillsAllowed: prescription.fillsAllowed,
+  daysSupply: prescription.daysSupply,
+  instructions: prescription.instructions,
+  notes: prescription.notes,
+  doNotFillBeforeDate: prescription.doNotFillBeforeDate,
+  diagnoses: prescription.diagnoseCodes
+});
 
 export function toPrescriptionFormData(
   prescription: Prescription,
